refactor(components/atom/backToTop): api tab
